recreation of a baffling error
Use F9 (if this is a J project) or manually load run.ijs to trigger the error of interest.

line 111 of wrapeval.ijs has an extraneous s, which triggers an error, but the error message does not adequately identify that location.

Enabling debugging reports an error on line 106 of wrapeval.ijs. That's a good hint, but it's also confusing. (The error should probably be reported on as occurring on line 93 of wrapeval.ijs, and in a sense it is - this is probably an error in the jqt debugger rather than the underlying 13!:n foreigns.)